Refactor `TokenService` from protocol into a single class.
